refactor(#41): 패키지명 수정 및 프로젝트 구조 재설계
